Separate timing and memory benchmark classes.
