+++ src/test/java/im/tox/tox4j/bench/MemoryReport.scala
@@ -1,0 +1,14 @@
+package im.tox.tox4j.bench
+
+import org.scalameter.api._
+
+abstract class MemoryReport extends PerformanceReportBase {
+
+  override val measurer = new Executor.Measurer.MemoryFootprint
+
+  // scalastyle:ignore
+  object memory extends Serializable {
+    def of(modulename: String): Scope = performance of (modulename + " (memory)") config defaultConfig
+  }
+
+}
