+++ src/test/java/im/tox/tox4j/bench/TimingReport.scala
@@ -1,0 +1,10 @@
+package im.tox.tox4j.bench
+
+abstract class TimingReport extends PerformanceReportBase {
+
+  // scalastyle:ignore
+  object timing extends Serializable {
+    def of(modulename: String): Scope = performance of modulename config defaultConfig
+  }
+
+}
