+++ src/test/java/im/tox/tox4j/core/bench/ToxCoreMemoryBench.scala
@@ -1,12 +1,9 @@
 package im.tox.tox4j.core.bench
 
-import im.tox.tox4j.Tox4jPerformanceReport
+import im.tox.tox4j.bench.MemoryReport
 import im.tox.tox4j.core.ToxCoreFactory
-import org.scalameter.api._
 
-class ToxCoreMemoryBench extends Tox4jPerformanceReport {
-
-  override val measurer = new Executor.Measurer.MemoryFootprint
+class ToxCoreMemoryBench extends MemoryReport {
 
   memory of "ToxCore" in {
